Significantly more bug fixes on WindowTextAreas. Fixed issue where WindowTextArea line numbers would not be drawn scissored. Fixed an issue with WindowTextArea's that would cause the highlighted line to stop being drawn once the text had gone out of frame.
Added additional EMC Events to handle catch when apps are reloaded and disabled as well as when the core is fully initialized.
Implemented logic to catch potentially broken EMCApps as they are being loaded into the core.
Broken apps are now displayed within the AppProblemsWindow (formerly known as IncompatibleWindowList).

General code formatting and clean up.